MNT #496 integrate into utils
